Repair #3 - works with referenced NuBuild projects
GetTargetPath target returns the TargetOutputs to VS/IDE when VS/IDE
requests the outputs of the projects for the reference dialog. We let
GetTargetPath to believe that NuBuild projects are .exe projects to let
them be able to be used as referenced projects. This has no effect
during real build, GetTargetPath is used by VS/IDE. But now we have to
filter out these .exe results from the TargetOutputs/ReferenceLibraries
list.