Different build workflow for VS and TFS/CI/MSBuild.
When you have a BuildingInsideVisualStudio property and you have to use
it, you know, sg is really broken with MSBuild... VS knows the build
order, we don't need MSBuild Target="Build" in this case, and CPP
projects can't be built this way. But on TFS/CI (ie. MSBuild.exe any
version) we have to use it.

This version finally works with mixed C#/C++ projects, with mixed
AnyCPU/x64 targets, with VS/TFS/CI/MSBuild.exe.